requirements.txt
requests
supabase
python-dotenv
pandas
linkedom
lxml
beautifulsoup4
pytest
